Setting Up SSH
Created a script to start the ssh-agent since eval ssh-agent doesn’t
work in fish.  Also created a function to start a connection to server
with tmux that will survive loss of internet.

I am not happy with the error handling of the tmux script.  I have
errors going to /dev/null, but that is because it spams the terminal if
the connection is lost.